Event Stack
Java Script works in a synchronous way.. how does it handle asynchronous events from browser?...such as click events...it does so by placing it in a event queue and executes that event queue after execution context in the scope chain are empty. This has been shown in the enclosed code.